Valida nota fiscal recebida e desconecta cliente no caso de falha

+++ lambda/invoices/invoiceImportUrlFunction.ts
@@ -1,7 +1,7 @@
 import { Context, S3Event, S3EventRecord } from 'aws-lambda'
 import { ApiGatewayManagementApi, DynamoDB, S3 } from 'aws-sdk'
 import * as AWSXRay from 'aws-xray-sdk'
-import { InvoiceTransactionRepository, InvoiceTransactionStatus } from '/opt/nodejs/invoiceTransaction'
+import { InvoiceTransaction, InvoiceTransactionRepository, InvoiceTransactionStatus } from '/opt/nodejs/invoiceTransaction'
 import { InvoiceWSService } from '/opt/nodejs/invoiceWSConnection'
 import { InvoiceFile, InvoiceRepository } from '/opt/nodejs/invoiceRepository'
 
@@ -37,7 +37,7 @@ async function processRecord(record: S3EventRecord): Promise<void> {
 
     try {
         const invoiceTransaction = await invoiceTransactionRepository.getInvoiceTransaction(key)
-
+        
         if (invoiceTransaction.transactionsStatus === InvoiceTransactionStatus.GENERATED) {
             await Promise.all([invoiceWSService.sendInvoiceStatus(key, invoiceTransaction.connectionId, InvoiceTransactionStatus.RECEIVED),
             invoiceTransactionRepository.updateInvoiceTransaction(key, InvoiceTransactionStatus.RECEIVED)])
@@ -55,29 +55,44 @@ async function processRecord(record: S3EventRecord): Promise<void> {
         const invoice = JSON.parse(object.Body!.toString('utf-8')) as InvoiceFile
         console.log(invoice)
 
-        const createInvoicePromise = invoiceRepository.create({
-            pk: `#invoice_${invoice.customerName}`,
-            sk: invoice.invoiceNumber,
-            ttl: 0,
-            totalValue: invoice.totalValue,
-            productId: invoice.productId,
-            quantity: invoice.quantity,
-            transactionId: key,
-            createdAt: Date.now()
-        })
+        const invoiceNumber = invoice.invoiceNumber
 
-        const deleteObjectPromise = s3Client.deleteObject({
-            Key: key,
-            Bucket: record.s3.bucket.name
-        }).promise()
+        if (invoiceNumber.length >= 5) {
+            const createInvoicePromise = invoiceRepository.create({
+                pk: `#invoice_${invoice.customerName}`,
+                sk: invoice.invoiceNumber,
+                ttl: 0,
+                totalValue: invoice.totalValue,
+                productId: invoice.productId,
+                quantity: invoice.quantity,
+                transactionId: key,
+                createdAt: Date.now()
+            })
 
-        const updateInvoicePromise = invoiceTransactionRepository.updateInvoiceTransaction(key, InvoiceTransactionStatus.PROCESSED)
+            const deleteObjectPromise = s3Client.deleteObject({
+                Key: key,
+                Bucket: record.s3.bucket.name
+            }).promise()
 
-        const sendStatusPromise = invoiceWSService.sendInvoiceStatus(key, invoiceTransaction.connectionId, InvoiceTransactionStatus.PROCESSED)
+            const updateInvoicePromise = invoiceTransactionRepository.updateInvoiceTransaction(key, InvoiceTransactionStatus.PROCESSED)
 
-        await Promise.all([createInvoicePromise, deleteObjectPromise, updateInvoicePromise, sendStatusPromise])
+            const sendStatusPromise = invoiceWSService.sendInvoiceStatus(key, invoiceTransaction.connectionId, InvoiceTransactionStatus.PROCESSED)
+
+            await Promise.all([createInvoicePromise, deleteObjectPromise, updateInvoicePromise, sendStatusPromise])
+        } else {
+            console.error(`Invoice import failed - non valid invoice number - TransactionId: ${key}`)
+
+            const sendStatusPromise = invoiceWSService.sendInvoiceStatus(key, invoiceTransaction.connectionId, InvoiceTransactionStatus.NON_VALID_INVOICE_NUMBER)
+
+            const updateInvoicePromise = invoiceTransactionRepository.updateInvoiceTransaction(key, InvoiceTransactionStatus.NON_VALID_INVOICE_NUMBER)
+
+            await Promise.all([sendStatusPromise, updateInvoicePromise])
+        }
+
     } catch (error) {
         console.log((<Error>error).message)
     }
 }
+
+
